make logging independent module
