+++ python/packages/autogen-agentchat/src/autogen_agentchat/teams/_base_team.py
@@ -5,12 +5,12 @@ from typing import List, Protocol
 from autogen_agentchat.agents._base_chat_agent import ChatMessage
 from autogen_core.application.logging import EVENT_LOGGER_NAME
 
-from ._utils import AgentChatLogHandler
+from .logging import ConsoleLogHandler, EVENT_LOGGER_NAME
 
-logger = logging.getLogger(EVENT_LOGGER_NAME + ".agentchatchat")
+logger = logging.getLogger(EVENT_LOGGER_NAME)
 logger.setLevel(logging.INFO)
-log_handler = AgentChatLogHandler(filename="log.jsonl")
-logger.handlers = [log_handler]
+console_handler = ConsoleLogHandler()
+logger.addHandler(console_handler)
 
 
 @dataclass
