+++ python/packages/autogen-agentchat/src/autogen_agentchat/teams/group_chat/_round_robin_group_chat.py
@@ -55,12 +55,7 @@ class RoundRobinGroupChat(BaseTeam):
 
     """
 
-    def __init__(
-        self,
-        participants: List[BaseChatAgent],
-        *,
-        tools: List[Tool] | None = None
-    ):
+    def __init__(self, participants: List[BaseChatAgent], *, tools: List[Tool] | None = None):
         if len(participants) == 0:
             raise ValueError("At least one participant is required.")
         if len(participants) != len(set(participant.name for participant in participants)):
@@ -75,8 +70,7 @@ class RoundRobinGroupChat(BaseTeam):
         def _factory() -> BaseChatAgentContainer:
             id = AgentInstantiationContext.current_agent_id()
             assert id == AgentId(type=agent.name, key=self._team_id)
-            container = BaseChatAgentContainer(
-                parent_topic_type, agent, tool_agent_type)
+            container = BaseChatAgentContainer(parent_topic_type, agent, tool_agent_type)
             assert container.id == id
             return container
 
@@ -95,8 +89,7 @@ class RoundRobinGroupChat(BaseTeam):
 
         # Register the tool agent.
         tool_agent_type = await ToolAgent.register(
-            runtime, "tool_agent", lambda: ToolAgent(
-                "Tool agent for round-robin group chat", self._tools)
+            runtime, "tool_agent", lambda: ToolAgent("Tool agent for round-robin group chat", self._tools)
         )
         # No subscriptions are needed for the tool agent, which will be called via direct messages.
 
@@ -109,8 +102,7 @@ class RoundRobinGroupChat(BaseTeam):
             topic_type = participant.name
             # Register the participant factory.
             await BaseChatAgentContainer.register(
-                runtime, type=agent_type, factory=self._create_factory(
-                    group_topic_type, participant, tool_agent_type)
+                runtime, type=agent_type, factory=self._create_factory(group_topic_type, participant, tool_agent_type)
             )
             # Add subscriptions for the participant.
             await runtime.add_subscription(TypeSubscription(topic_type=topic_type, agent_type=agent_type))
@@ -132,16 +124,13 @@ class RoundRobinGroupChat(BaseTeam):
         )
         # Add subscriptions for the group chat manager.
         await runtime.add_subscription(
-            TypeSubscription(topic_type=group_chat_manager_topic_type,
-                             agent_type=group_chat_manager_agent_type.type)
+            TypeSubscription(topic_type=group_chat_manager_topic_type, agent_type=group_chat_manager_agent_type.type)
         )
         await runtime.add_subscription(
-            TypeSubscription(topic_type=group_topic_type,
-                             agent_type=group_chat_manager_agent_type.type)
+            TypeSubscription(topic_type=group_topic_type, agent_type=group_chat_manager_agent_type.type)
         )
         await runtime.add_subscription(
-            TypeSubscription(topic_type=team_topic_type,
-                             agent_type=group_chat_manager_agent_type.type)
+            TypeSubscription(topic_type=team_topic_type, agent_type=group_chat_manager_agent_type.type)
         )
 
         group_chat_messages: List[ChatMessage] = []
@@ -156,8 +145,7 @@ class RoundRobinGroupChat(BaseTeam):
             type="collect_group_chat_messages",
             closure=collect_group_chat_messages,
             subscriptions=lambda: [
-                TypeSubscription(topic_type=group_topic_type,
-                                 agent_type="collect_group_chat_messages")
+                TypeSubscription(topic_type=group_topic_type, agent_type="collect_group_chat_messages")
             ],
         )
 
@@ -166,11 +154,9 @@ class RoundRobinGroupChat(BaseTeam):
 
         # Run the team by publishing the task to the team topic and then requesting the result.
         team_topic_id = TopicId(type=team_topic_type, source=self._team_id)
-        group_chat_manager_topic_id = TopicId(
-            type=group_chat_manager_topic_type, source=self._team_id)
+        group_chat_manager_topic_id = TopicId(type=group_chat_manager_topic_type, source=self._team_id)
         await runtime.publish_message(
-            ContentPublishEvent(agent_message=TextMessage(
-                content=task, source="user")),
+            ContentPublishEvent(agent_message=TextMessage(content=task, source="user")),
             topic_id=team_topic_id,
         )
         await runtime.publish_message(ContentRequestEvent(), topic_id=group_chat_manager_topic_id)
